+++ app/_components/common/header/_component/Title.tsx
@@ -1,4 +1,4 @@
-import style from './icon.module.css'
+import style from './icon.module.scss'
 
 export default function Title({ title }: { title: string }) {
   return <div className={style.container}>{title}</div>
