+++ app/_components/postPreview/PostPreview.tsx
@@ -1,4 +1,4 @@
-import style from './postPreview.module.css'
+import style from './postPreview.module.scss'
 
 import Link from '@/node_modules/next/link'
 import Image from '@/node_modules/next/image'
