+++ app/_components/postPreview/_component/Post.tsx
@@ -1,4 +1,4 @@
-import style from './post.module.css'
+import style from './post.module.scss'
 import Image from '@/node_modules/next/image'
 
 import Like from '@/public/assets/like.svg'
