+++ app/_components/postPreview/_component/post.module.scss
