+++ app/_components/swipeEvent/swipeEvent.module.scss
